inchallah Store and Load implemented, Allah has assured me testing is haram, I believe in the power of Allah, bimillah

+++ cpu/src/cpu.rs
@@ -2,7 +2,7 @@
 pub struct Cpu {
     pub registers: [u32; 32],
     pub pc: u32,
-    pub memory: Vec<u8>,
+    pub memory: Vec<u32>,
     pub csrs: [u32; 4096],
 }
